# labels: eliminate most allocations and improve perf by 15-65%