--- a/labels_test.go
+++ b/labels_test.go
@@ -6,38 +6,25 @@ import (
 )
 
 func TestCompareDomainName(t *testing.T) {
-	s1 := "www.miek.nl."
-	s2 := "miek.nl."
-	s3 := "www.bla.nl."
-	s4 := "nl.www.bla."
-	s5 := "nl."
-	s6 := "miek.nl."
-
-	if CompareDomainName(s1, s2) != 2 {
-		t.Errorf("%s with %s should be %d", s1, s2, 2)
+	tests := []struct {
+		s1, s2   string
+		expected int
+	}{
+		{"www.miek.nl.", "miek.nl.", 2},
+		{"miek.nl.", "www.bla.nl.", 1},
+		{"www.bla.nl.", "nl.www.bla.", 0},
+		{"www.miek.nl.", "nl.", 1},
+		{"www.miek.nl.", "miek.nl.", 2},
+		{"www.miek.nl.", ".", 0},
+		{".", ".", 0},
+		{"test.com.", "TEST.COM.", 2},
+		{"a.b.c.d.e.f.", "a.b.c.d.e.", 0},
+		{"a.b.c.d.e.", "a.b.c.d.e.", 5},
 	}
-	if CompareDomainName(s1, s3) != 1 {
-		t.Errorf("%s with %s should be %d", s1, s3, 1)
-	}
-	if CompareDomainName(s3, s4) != 0 {
-		t.Errorf("%s with %s should be %d", s3, s4, 0)
-	}
-	// Non qualified tests
-	if CompareDomainName(s1, s5) != 1 {
-		t.Errorf("%s with %s should be %d", s1, s5, 1)
-	}
-	if CompareDomainName(s1, s6) != 2 {
-		t.Errorf("%s with %s should be %d", s1, s5, 2)
-	}
-
-	if CompareDomainName(s1, ".") != 0 {
-		t.Errorf("%s with %s should be %d", s1, s5, 0)
-	}
-	if CompareDomainName(".", ".") != 0 {
-		t.Errorf("%s with %s should be %d", ".", ".", 0)
-	}
-	if CompareDomainName("test.com.", "TEST.COM.") != 2 {
-		t.Errorf("test.com. and TEST.COM. should be an exact match")
+	for _, x := range tests {
+		if i := CompareDomainName(x.s1, x.s2); i != x.expected {
+			t.Errorf("%s with %s should be %d got: %d", x.s1, x.s2, x.expected, i)
+		}
 	}
 }
 
@@ -106,6 +93,9 @@ func TestPrevLabel(t *testing.T) {
 
 		{"www.miek.nl.", 3}: 0,
 		{"www.miek.nl", 3}:  0,
+
+		{"a.b.c.", 1}: 4,
+		{"a.b.c", 1}:  4,
 	}
 
 	// make sure we are safe when the label  begins with a possibly escaped '.'